fix(BulkDesynth): widen DesynthUnlockQuest constant to uint
65688 exceeds ushort.MaxValue. QuestManager.IsQuestComplete has a uint
overload that masks to the native ushort id (65688 & 0xFFFF = 152), so
passing it as uint resolves to the correct quest with no truncation
surprises.

+++ FF14/BulkDesynth/Services/InventoryScanner.cs
@@ -18,8 +18,9 @@ namespace BulkDesynth.Services;
 public sealed class InventoryScanner
 {
     /// <summary>Quest row ID that unlocks desynthesis. Sourced from AutoRetainer's
-    /// reference implementation; the quest is "Gone to Pieces" / equivalent.</summary>
-    private const ushort DesynthUnlockQuest = 65688;
+    /// reference implementation; the quest is "Gone to Pieces" / equivalent.
+    /// Stored as uint because the row ID exceeds ushort.MaxValue (65535).</summary>
+    private const uint DesynthUnlockQuest = 65688;
 
     /// <summary>Minimum DoH class level required to use desynthesis at all.</summary>
     private const byte MinDohLevelForDesynth = 30;
